Add Field Mapping Support for GeoPointV2
This commit adds an abstraction layer to the GeoPointFieldMapper for supporting the new GeoPointField type in Lucene 5.4. The abstraction supports backward compatibility with Legacy geo_point indexes.
