+++ core/src/main/java/org/elasticsearch/index/mapper/MapperBuilders.java
@@ -19,11 +19,13 @@
 
 package org.elasticsearch.index.mapper;
 
+import org.elasticsearch.Version;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.index.mapper.core.*;
+import org.elasticsearch.index.mapper.geo.BaseGeoPointFieldMapper;
 import org.elasticsearch.index.mapper.geo.GeoPointFieldMapper;
+import org.elasticsearch.index.mapper.geo.GeoPointFieldMapperLegacy;
 import org.elasticsearch.index.mapper.geo.GeoShapeFieldMapper;
-import org.elasticsearch.index.mapper.internal.*;
 import org.elasticsearch.index.mapper.ip.IpFieldMapper;
 import org.elasticsearch.index.mapper.object.ObjectMapper;
 import org.elasticsearch.index.mapper.object.RootObjectMapper;
@@ -92,8 +94,11 @@ public final class MapperBuilders {
         return new DoubleFieldMapper.Builder(name);
     }
 
-    public static GeoPointFieldMapper.Builder geoPointField(String name) {
-        return new GeoPointFieldMapper.Builder(name);
+    public static BaseGeoPointFieldMapper.Builder geoPointField(String name, Version version) {
+        // TODO switch to version.onOrBefore(Version.V_2_1_0) once GeoPointField V2 is fully merged
+        return (version.onOrBefore(Version.V_2_2_0) ?
+                new GeoPointFieldMapperLegacy.Builder(name) :
+                new GeoPointFieldMapper.Builder(name));
     }
 
     public static GeoShapeFieldMapper.Builder geoShapeField(String name) {
